archive/health_test_uploader/requirements.txt
fastapi
uvicorn
pytest
python-multipart
python-dotenv
openai
PyPDF2